Refactoring, fix theme
